// src/app/generated/models/analysis-detail.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AnalysisMode } from '../models/analysis-mode';
import { Candidate } from '../models/candidate';
import { DecisionPath } from '../models/decision-path';
import { ExtractedSymptom } from '../models/extracted-symptom';
import { Thresholds } from '../models/thresholds';
export interface AnalysisDetail {
  asrConfidence?: number | null;
  candidates: Array<Candidate>;
  caseId: string;
  decisionPath: DecisionPath;
  errorCode?: string | null;
  hasGradcam?: boolean;
  latencyMs?: number;
  margin?: number | null;
  mode: AnalysisMode;
  symptoms: Array<ExtractedSymptom>;
  thresholds: Thresholds;
  top1Confidence?: number | null;
  top2Confidence?: number | null;
  transcriptBn?: string | null;
  unmappedLabels?: Array<string>;
  visionModelId?: string | null;
  visionModelVersion?: string | null;
}